Fix: Uninitialized variable access in Cycles
Shadow terminator tricks might access uninitialized value of `wo`,
which happens in cases BSDF evaluation returns LABEL_NONE (i.e. when
the incident ray is in the lower hemisphere).

It is likely harmless from the render result perspective, as the `eval`
does not seem to be used. Also, since the `eval` is initialized to 0
it would need to be aa case when undefined `wo` leads to NAN or INF
returned by the shadow terminator offsets. Still good to solve this issue
to allow use of debugging tools like Valgrind.

Pull Request: https://projects.blender.org/blender/blender/pulls/149932