Refactor: Remove primitive_type from HIP-RT payload
Move handling related to packed segment type to the conversion of
hiprtHit to Intersection. The benefits of doing so include:

- Reduces the size of the payload
- Solves confusion with point clouds: the previous code that was
  handling isect->type applied it for all non-triangle primitives
  while comment was only referring to curves.
- Makes payload look more like the one for Metal-RT, making it
  easier to eventually unify the payload and code.

There are a bit of extra fetching involved into intersection after
this change. For now it is not a concern, as the code is planned to
be more heavily refactored.